requirements.txt
torch
torchvision
ultralytics
numpy
opencv-python
Pillow
pyyaml
coremltools
pyyaml 
Pillow 
gradio 
